+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.5.1] - 2023-06-29
+### Fixed
+- Regular expression for `{{ set }}` tag.
+- Multiline code for `set`, `include`, `for`
+
 ## [0.5.0] - 2023-06-18
 ### Added
 - Allow to add arbitrary JavaScript expressions with `{{> [js code] }}`.
@@ -52,6 +57,7 @@ First version
 [#4]: https://github.com/oscarotero/vento/issues/4
 [#5]: https://github.com/oscarotero/vento/issues/5
 
+[0.5.1]: https://github.com/oscarotero/vento/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/oscarotero/vento/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/oscarotero/vento/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/oscarotero/vento/compare/v0.3.0...v0.3.1